// src/types/constants.ts
// src/types/constants.ts
export interface ProcessStep {
    title: string;
    description: string;
    time: string;
    icon: string;
  }
  
  export interface Solution {
    icon: string;
    title: string;
    description: string;
  }
  
  export interface SelectOption {
    value: string;
    label: string;
  }
  
  export interface TrustElement {
    title: string;
    icon: string;
  }
  
  export interface CTAFeature {
    icon: string;
    title: string;
  }